requirements.txt
aiogram>=3.24,<4.0
aiohttp-socks>=0.9,<1.0
google-genai>=1.30,<2.0
Pillow>=10.4,<12.0
python-dotenv>=1.0,<2.0
